Improve compiling
Use cross compiler instead of normal G++
Improve compiler options

+++ kernel/include/thor.hpp
@@ -1,11 +1,9 @@
 #ifndef THOR_H
 #define THOR_H
 
-#include <new>
-
 #include "memory.hpp"
 
-void* operator new (size_t size);
+void* operator new (uint64_t size);
 void operator delete (void *p);
 
 #endif
